% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_reg.R
\name{multi_predict._fishnet}
\alias{multi_predict._fishnet}
\title{Model predictions across many sub-models}
\usage{
\method{multi_predict}{`_fishnet`}(object, new_data, type = NULL, penalty = NULL, ...)
}
\arguments{
\item{penalty}{A numeric vector of penalty values.}
}
\description{
Model predictions across many sub-models
}
\keyword{internal}
