% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seniors}
\alias{seniors}
\title{Alcohol, Cigarette, and Marijuana Use for High School Seniors}
\source{
Agresti, A (2007). \emph{An Introduction to Categorical Data Analysis}.
}
\value{
\item{seniors}{a tibble}
}
\description{
Alcohol, Cigarette, and Marijuana Use for High School Seniors
}
\details{
Data are from Table 7.3 of Agresti (2007). The first three columns
make up data from a 3-way contingency table.
}
\examples{
data(seniors)
str(seniors)
}
\keyword{datasets}
